Move ErrorMessageResponse from core to web module
