+++ molgenis-core-ui/src/main/java/org/molgenis/ui/style/StyleController.java
@@ -1,7 +1,7 @@
 package org.molgenis.ui.style;
 
 import org.apache.commons.io.IOUtils;
-import org.molgenis.util.ErrorMessageResponse;
+import org.molgenis.web.ErrorMessageResponse;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
