+++ molgenis-security/pom.xml
@@ -21,6 +21,11 @@
     <dependencies>
         <dependency>
             <groupId>org.molgenis</groupId>
+            <artifactId>molgenis-web</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.molgenis</groupId>
             <artifactId>molgenis-security-core</artifactId>
             <version>${project.version}</version>
         </dependency>
